feat: :boom: class file seperated to the utils folder
