+++ Back-End/Jonas Node+MongoDB Course/Nodejs + ExpressJS/Express-API/controllers/tourController.js
@@ -1,59 +1,5 @@
 const Tour = require('../models/tourModel.js');
-
-//=============== Class ================
-class APIfeatures {
-  constructor(query, queryString) {
-    this.query = query;
-    this.queryString = queryString;
-  }
-
-  filter() {
-    const queryObj = { ...this.queryString };
-    const excludedFields = ['page', 'limit', 'sort', 'fields'];
-
-    //Now we will loop over the excludedFields and remove the element from queryObject if it's present
-    excludedFields.forEach(el => {
-      delete queryObj[el];
-    });
-
-    //we get req.query as original and then we remove the fields added in excluded and get new object as queryObj which we use for quering
-    console.log(this.queryString, queryObj); //Used for filetering in tours
-
-    this.query = this.query.find(queryObj);
-
-    return this;
-  }
-
-  sort() {
-    if (this.queryString.sort) {
-      this.query = this.query.sort(this.queryString.sort);
-    } else {
-      this.query = this.query.sort('-createdAt;');
-    }
-
-    return this;
-  }
-
-  limitFields() {
-    if (this.queryString.fields) {
-      const field = this.queryString.fields.split(',').join(' ');
-      this.query = this.query.select(field);
-    } else {
-      this.query = this.query.select('-__v');
-    }
-
-    return this;
-  }
-
-  pagination() {
-    const page = this.queryString.page * 1 || 1;
-    const limit = this.queryString.limit * 1 || 100;
-    const skip = (page - 1) * limit;
-    this.query = this.query.skip(skip).limit(limit);
-
-    return this;
-  }
-}
+const APIfeatures = require('../utils/apiFeatures.js');
 
 //================ Get all tours =========================
 exports.getAllTours = async (req, res) => {
